refactor: 모듈 구조 및 TypeOrm 연결 방식 변경
auth, oauth, user 모듈 -> auth(/, /oauth), user 모듈

TypeOrmModule 이용해 repository pattern으로 user DB 구현
