+++ server/package.json
@@ -33,6 +33,8 @@
 		"axios": "^1.1.3",
 		"cache-manager": "^5.1.3",
 		"cache-manager-ioredis": "^2.1.0",
+		"class-transformer": "^0.5.1",
+		"class-validator": "^0.13.2",
 		"connect-redis": "^6.1.3",
 		"cross-env": "^7.0.3",
 		"dotenv": "^16.0.3",
