+++ server/src/app.module.ts
@@ -1,29 +1,20 @@
-import { Module, CacheModule } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { UserModule } from './user/user.module';
-import { DatabaseModule } from './database/database.module';
-import { AuthModule } from './auth/auth.module';
-import { OauthService } from './oauth/oauth.service';
-import { OauthModule } from './oauth/oauth.module';
 import { HttpModule } from '@nestjs/axios';
-import { AuthService } from './auth/auth.service';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { typeORMConfig } from './configs/typeorm.config';
+import { AuthModule } from './auth/auth.module';
 
 @Module({
   imports: [
+    TypeOrmModule.forRoot(typeORMConfig),
     UserModule,
-    DatabaseModule,
-    AuthModule,
-    OauthModule,
     HttpModule,
-    CacheModule.register({
-      isGlobal: true,
-      // store: redisStore,
-      // host: process.env.REDIS_HOST,
-      // port: process.env.REDIS_PORT,
-    }),
+    AuthModule,
   ],
   controllers: [AppController],
-  providers: [AppService, OauthService, AuthService],
+  providers: [AppService],
 })
 export class AppModule {}
