--- server/src/auth/auth.dto.ts
@@ -1,4 +1,0 @@
-export class AuthDto {
-  oauthType: string;
-  email: string;
-}
