+++ server/src/auth/auth.module.ts
@@ -1,14 +1,10 @@
-import { CacheModule, Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
+import { Module } from '@nestjs/common';
 import { UserModule } from 'src/user/user.module';
 import { AuthController } from './auth.controller';
+import { AuthService } from './auth.service';
+
 @Module({
-  imports: [
-    UserModule,
-    CacheModule.register({
-      isGlobal: true,
-    }),
-  ],
+  imports: [UserModule],
   controllers: [AuthController],
   providers: [AuthService],
 })
