+++ server/src/configs/typeorm.config.ts
@@ -1,0 +1,22 @@
+import * as dotenv from 'dotenv';
+import * as path from 'path';
+import { TypeOrmModuleOptions } from '@nestjs/typeorm';
+
+dotenv.config({
+  path: path.resolve(
+    process.env.NODE_ENV === 'production'
+      ? '.production.env'
+      : '.development.env',
+  ),
+});
+
+export const typeORMConfig: TypeOrmModuleOptions = {
+  type: 'mysql',
+  host: process.env.DATABASE_HOST,
+  port: 3306,
+  username: process.env.DATABASE_USERNAME,
+  password: process.env.DATABASE_PASSWORD,
+  database: process.env.DATABASE_NAME,
+  entities: ['../**/*.entity.ts'],
+  synchronize: true,
+};
