--- server/src/database/database.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import { databaseProviders } from './database.provider';
-
-@Module({
-  providers: [...databaseProviders],
-  exports: [...databaseProviders],
-})
-export class DatabaseModule {}
