--- server/src/database/database.provider.ts
@@ -1,31 +1,0 @@
-import { DataSource } from 'typeorm';
-import * as path from 'path';
-import * as dotenv from 'dotenv';
-import { User } from 'src/user/user.entity';
-dotenv.config({
-  path: path.resolve(
-    process.env.NODE_ENV === 'production'
-      ? '.production.env'
-      : '.development.env',
-  ),
-});
-
-export const databaseProviders = [
-  {
-    provide: 'DATA_SOURCE',
-    useFactory: async () => {
-      const dataSource = new DataSource({
-        type: 'mysql',
-        host: process.env.DATABASE_HOST,
-        port: 3306,
-        username: process.env.DATABASE_USERNAME,
-        password: process.env.DATABASE_PASSWORD,
-        database: process.env.DATABASE_NAME,
-        entities: [User],
-        synchronize: true,
-      });
-
-      return dataSource.initialize();
-    },
-  },
-];
