--- server/src/oauth/oauth.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { OauthController } from './oauth.controller';
-
-describe('OauthController', () => {
-  let controller: OauthController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [OauthController],
-    }).compile();
-
-    controller = module.get<OauthController>(OauthController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
