--- server/src/user/user.providers.ts
@@ -1,10 +1,0 @@
-import { DataSource } from 'typeorm';
-import { User } from './user.entity';
-
-export const userProviders = [
-  {
-    provide: 'USER_REPOSITORY',
-    useFactory: (dataSource: DataSource) => dataSource.getRepository(User),
-    inject: ['DATA_SOURCE'],
-  },
-];
